Allow using Template directly
I needed to change the delimiter but couldn't because the fields on
Template are unexported and so is the parse function. I have added a
constructor for Template, named NewTemplate, that sets all of its
fields. Then I exported Parse so that someone could create a template
directly and then compile it.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>